Added Ability System files
Modified Ability.cs which is an abstract class for an "Ability" ScriptableObject, to have an activeTime variable.
Added RangedAbility.cs which inherits from the Ability class and defines parameters for a ranged ability (may need some tweaking in projectile logic).
Added AbilityHolder.cs script which can be attached to an GameObject for it to be able to perform an "Ability" on a certain keypress.
Added an AbilityHolder.cs to Player that uses the "Fireball" ScriptableObject on the spacebar.
Added rest of dungeon assets.