Added Melee Ability
Added MeleeAbility.cs which inherits from the Ability class and defines the behavior of a basic melee attack. (Currently the key for Melee is set to "F')
Added Melee.prefab which is created when MeleeAbility is used and destroyed after MeleeAbility is ended.
Currently there is no code for dealing damage.