Added Phase Ability
Added a phase ability that makes the player become "invisible" by pressing "c".

Modified Enemy.cs script: added an if statement to check if the player is "invisible" before chasing the player.

Need to discuss if we want the player to be immune to all damages or just become "invisible" to the enemies.