Tidy up damage system
Now the damage of both melee and ranged ability are controlled by the gameobject created in the scripts folder. The "damage" value is passed to "damage" in Bullet and AbilityDamage
Bullet and AbilityDamage is still showing a "damage" variable waiting for a value. Do not use it, and hopefully they won't make any effect when given a value.